BED-4937 moved user modification forms/dialogs to shared ui
