+++ packages/javascript/bh-shared-ui/src/components/UpdateUserDialog/UpdateUserDialog.tsx
@@ -16,8 +16,8 @@
 
 import { Dialog, DialogTitle } from '@mui/material';
 import React from 'react';
-import { UpdatedUser } from 'src/ducks/auth/types';
-import UpdateUserForm from './UpdateUserForm';
+import UpdateUserForm from '../UpdateUserForm';
+import {UpdatedUser} from "../../ducks";
 
 const UpdateUserDialog: React.FC<{
     open: boolean;
