+++ packages/javascript/bh-shared-ui/src/components/UpdateUserDialog/index.ts
@@ -1,0 +1,4 @@
+export { default } from './UpdateUserDialog'
+
+export * from './UpdateUserDialog'
+
