+++ packages/javascript/bh-shared-ui/src/components/UpdateUserForm/index.ts
@@ -1,0 +1,3 @@
+export * from './UpdateUserForm';
+
+export { default } from "./UpdateUserForm";
