+++ packages/javascript/bh-shared-ui/src/index.ts
@@ -69,3 +69,5 @@ export * from './views';
 export * from './store';
 
 export * from './mocks';
+
+export * from './ducks';
